# apiGateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway для микросервисного приложения вишлистов",
        "title": "Wishlist API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8095",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Аутентификация пользователя и получение JWT токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Учетные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "schema": {
                            "$ref": "#/definitions/handlers.AuthDTO"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Создает нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешная регистрация",
                        "schema": {
                            "$ref": "#/definitions/handlers.AuthDTO"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "409": {
                        "description": "Пользователь уже существует",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
        "/auth/self": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает данные аутентифицированного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Информация о текущем пользователе",
                "responses": {
                    "200": {
                        "description": "Информация о пользователе",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserInfoDTO"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AuthDTO": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "description": "ExpiresAt Время окончания токена (Unix timestamp)\nformat: date-time\nexample: 1700000000",
                    "type": "integer"
                },
                "refresh_token": {
                    "description": "Токен для обновления доступа\nexample: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "type": "string"
                },
                "token": {
                    "description": "JWT токен доступа\nexample: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID UUID пользователя\nformat: uuid\nexample: 123e4567-e89b-12d3-a456-426614174000",
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error содержит строку с расшифровкой того что пошло не так\nminLength: 1\nmaxLength: 100\nexample: item not found",
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email пользователя\nrequired: true\nformat: email\nexample: user@example.com",
                    "type": "string"
                },
                "password": {
                    "description": "Пароль пользователя (минимум 4 символа)\nrequired: true\nminLength: 4\nexample: password123",
                    "type": "string"
                }
            }
        },
        "handlers.RegisterRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email пользователя\nrequired: true\nformat: email\nexample: user@example.com",
                    "type": "string"
                },
                "name": {
                    "description": "Имя, ник пользователя\nrequired: true\nexample: Иван",
                    "type": "string"
                },
                "password": {
                    "description": "Пароль пользователя (минимум 4 символа)\nrequired: true\nminLength: 4\nexample: password123",
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "handlers.UserInfoDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email пользователя\nrequired: true\nformat: email\nexample: user@example.com",
                    "type": "string"
                },
                "name": {
                    "description": "Name Имя, ник пользователя\nrequired: true\nexample: Иван",
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID UUID пользователя\nformat: uuid\nexample: 123e4567-e89b-12d3-a456-426614174000",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}